refactor: View를 하나로 통합
