--- Diary/Scene/Detail/View/DetailView.swift
@@ -1,19 +1,0 @@
-//
-//  DetailView.swift
-//  Diary
-//
-//  Created by Taeangel, Quokka on 2022/06/22.
-//
-
-import UIKit
-
-final class DetailView: WriteView {
-  func updateTextView(diary: Diary) {
-    guard let title = diary.title,
-          let body = diary.content else {
-      return
-    }
-    let content = title + "\n" + body
-    textView.text = content
-  }
-}
