--- Diary/Scene/Write/View/WriteView.swift
@@ -1,10 +1,0 @@
-//
-//  WriteView.swift
-//  Diary
-//
-//  Created by Taeangel, Quokka on 2022/06/15.
-//
-
-import UIKit
-
-class WriteView: DiaryBaseView {}
